Order duplicates resolved

+++ src/main/java/avansivh11/dehartigesupermarkt/model/order/BaseOrder.java
@@ -16,5 +16,5 @@ public abstract class BaseOrder {
 	@GeneratedValue
 	private Long id;
 
-	public abstract int price();
+	public abstract double price();
 }
